Added a light and shading to planets
I made shaders for the planets and parented a point light to the player so you can fly around the planets and have some dynamic lighting.

This is not really what we would want but it was the easiest way of testing shaders. In a real level I would think we'd have a sun, some bounce lights that affect the player and then perhaps some weapons cast light and shadows as well.

Anyway, a fun little test.