Tweaked Player stats and Starfield
Feel free to reject this commit, but I wanted to play a bit with the flying and the starfield. I moved some stars to the players plane and also have more movement in the deep BG stars (less realistic I know but more snappy feeling while flying). I balanced the thrust and speeds a bit and I adjusted the pivot point of the ship. Maybe give it a try and see how it feels.