Added Dash and Power Dash
While thursting, if one presses the (-) key they get a small burst of speed. If you hold it for longer than 1 second you get a bigger longer burst.

The fiction is that you are dumping a small amount of hyperfuel into your impulse drive. So theoretically it has a fuel cost perhaps.